add pages directory and refactor Home
