+++ frontend/src/App.tsx
@@ -2,7 +2,7 @@ import './App.css'
 import axios from "axios";
 import {useEffect, useState} from "react";
 import {Movie} from "./Movie.ts";
-import Home from "./components/Home.tsx";
+import Home from "./pages/Home.tsx";
 import {Route, Routes} from "react-router-dom";
